Fixed loading of shift entry and timesheet templates. Got shift entry template to display, still not functional for punch in/out

+++ src/main/java/com/buffettinc/hrms/service/time/ShiftEntryService.java
@@ -81,5 +81,5 @@ public interface ShiftEntryService {
      * @param shiftID The ID of the shift entry.
      * @return The duration of the shift in hours.
      */
-    long getShiftDurationInHours(Long shiftID);
+    double getShiftDurationInHours(Long shiftID);
 }
